tests/btreetest.cc: Return filesize as off_t to avoid MSVC warning about truncating from 64 to 32 bits (the filesize will never be more than a few K here, so this is essentially a cosmetic issue).
git-svn-id: file:///home/xapian-svn/xapian/trunk@8520 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/tests/btreetest.cc
@@ -60,7 +60,7 @@ static void make_dir(const string & filename)
 }
 
 /// Get the size of the given file in bytes.
-static int get_filesize(const string &filename)
+static off_t get_filesize(const string &filename)
 {
     struct stat buf;
     int result = stat(filename, &buf);
